requirements.txt
# 基于Level-2数据的A股大资金成本与意图分析系统
# Python依赖包

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 数据源
akshare>=1.12.0
tushare>=1.4.0

# 配置文件
pyyaml>=6.0

# 数据库
sqlalchemy>=2.0.0

# 日志
loguru>=0.7.0

# 类型检查
typing-extensions>=4.5.0

# 日期时间
python-dateutil>=2.8.0

# 进度条
tqdm>=4.65.0

# 数据验证
pydantic>=2.0.0

# HTTP请求
requests>=2.31.0

# 异步处理
asyncio>=3.4.3

# 测试框架
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
